% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rle.R
\name{c_rle}
\alias{c_rle}
\title{Compute the Run Length Encoding}
\usage{
c_rle(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Compute the run length encoding. This recreates the \code{\link[=rle]{rle()}} function, but written in C.
}
\examples{
c_rle(c(1, 1, 3, 3, 3, 3, 2, 2, 2))

}
