% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crle-package.R
\docType{package}
\name{crle-package}
\alias{crle}
\alias{crle-package}
\title{crle: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/crle}
  \item Report bugs at \url{https://github.com/EllaKaye/crle/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
