% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crle-package.R
\docType{package}
\name{crle-package}
\alias{crle}
\alias{crle-package}
\title{crle: C backend for rle}
\description{
A demonstration package including C code
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EllaKaye/crle}
  \item Report bugs at \url{https://github.com/EllaKaye/crle/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.m.kaye@warwick.ac.uk}

}
\keyword{internal}
